Added confirmation message for transaction deletion

+++ client/coinflow/src/components/TransactionEditorComponent.vue
@@ -226,8 +226,10 @@ export default {
       }
     },
     async onDeleteTransaction() {
-      await fetcher.deleteTransaction(this.transaction)
-      await this.$emit('transactionDeleted')
+      if (confirm(`Really delete transaction '${this.transaction.description}'?`)) {
+        await fetcher.deleteTransaction(this.transaction)
+        await this.$emit('transactionDeleted')
+      }
     }
   },
   computed: {
